+++ lib/imcodec.dart
@@ -2,6 +2,7 @@
 library;
 
 export 'src/codecs/jpeg_encoder.dart' show JpegChroma;
+export 'src/codecs/tiff_codec.dart' show TiffCompression;
 export 'src/decoder.dart';
 export 'src/encoder.dart';
 export 'src/image.dart';